pyproject.toml
[tool.poetry]
name = "mixs_subset_examples_first"
version = "0.0.0.post3.dev0+f188a04"
description = "A subset of the MIxS specification that's self-documenting, DataHarmonizer compatible. Comes with valid and invalid data examples."
authors = ["Mark Andrew Miller <MAM@lbl.gov>"]
license = "MIT"
readme = "README.md"
include = ["README.md", "src/mixs_subset_examples_first/schema", "project"]

[tool.poetry.dependencies]
python = "^3.9"
linkml-runtime = "^1.1.24"
oaklib = "^0.5.0"
xmltodict = "^0.13.0"

[tool.poetry-dynamic-versioning]
enable = false
vcs = "git"
style = "pep440"

[tool.poetry.dev-dependencies]
linkml = "^1.3.5"
mkdocs-material = "^8.2.8"
mkdocs-mermaid2-plugin = "^0.6.0"
schemasheets = "^0.1.14"
check-jsonschema = "^0.22.0"
click_log = "0.4.0"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry.extras]
docs = ["linkml", "mkdocs-material"]

[tool.poetry.scripts]
dh-json2linkml = 'src.mixs_subset_examples_first.datamodel.dh_json2linkml:update_json'
linkml-json2dh = 'src.mixs_subset_examples_first.datamodel.linkml_json2dh:extract_lists'

combine_any_tsvs = 'src.mixs_subset_examples_first.datamodel.combine_any_tsvs:combine_tsvs'
compare_headers = 'src.mixs_subset_examples_first.datamodel.compare_headers:cli'
find_contradictions = 'src.mixs_subset_examples_first.datamodel.find_contradictions:main'
tsv_column_alignment = 'src.mixs_subset_examples_first.datamodel.tsv_column_alignment:cli'
xlsx_tab_to_tsv = 'src.mixs_subset_examples_first.datamodel.xlsx_tab_to_tsv:cli'
drop_then_remove_dupes = 'src.mixs_subset_examples_first.datamodel.drop_then_remove_dupes:main'

filter_column = 'src.mixs_subset_examples_first.datamodel.filter_column:cli'
contradiction_score_reports = 'src.mixs_subset_examples_first.datamodel.contradiction_score_reports:main'
headers_as_lines = 'src.mixs_subset_examples_first.datamodel.headers_two:print_headers'
extract_unique_column_vals = 'src.mixs_subset_examples_first.datamodel.unique_two:extract_unique_values'
combine_checklists_and_env_packages = 'src.mixs_subset_examples_first.datamodel.combine_checklists_and_env_packages:generate_combined_dataframe'
combine_same_col_schemasheets = 'src.mixs_subset_examples_first.datamodel.combine_same_col_schemasheets:combine_tsvs'
melt_tsv = 'src.mixs_subset_examples_first.datamodel.melt_tsv:melt_tsv'



